+++ docs/changelog.rst
@@ -16,6 +16,8 @@ Changed
 
 
 * Incident class accepts as a priority only a Priority enum
+* Incident class must be instantiated with a step number
+* Step number in Incident class must be an integer, null is not allowed anymore
 
 Removed
 ^^^^^^^